Moved share session logs to settings view
